refactor(db): Create database service for testability
### Problem Statement

The database layer was difficult to test without really having the
ability to hang a mock from an object as it was all functions.

### Solution

Encapsulate database read and write into a class for future test-
ability. This is in preparation for refactoring the server
so it can be unit tested.

### How has this been tested?

Existing tests were refactored and cover much of the code that has
been modified.