patch(database.db): remove checked in test file
### Problem Statement
The database.db file got checked in but that was not my intention. Another
issue was the fact that when using `reset-db` to do initial db migration
when the db file is not there it was crashing.

### Solution
I am sneaking in a small fix here adding the `-f` flag to the `rm` for
`reset-db`. Also, removing database.db from git and fixing the gitignore

### How has this been tested?
local testing with `git add .` and `yarn reset-db`