test(ingest_block): Add input validation testing
### Problem Statement

We did not have any test coverage for the server.

### Solution

This starts to show how we can use mocking to mock the DB layer and the
constructor injection pattern to test various parts of the server
code.

Eventually the blockchain provider must also be mocked and injected for
further testing. In the end `server.test.ts` should be able to be
completely unit tested without much trouble. We may want to create
a service layer for blockchain data similar to the database layer, right
now much of that logic is encapsulated into the route itself.

### How has this been tested?

Unit tests included.